@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": "^5.4|^7.0",
+        "php": "^7.2",
         "ext-json": "*",
         "ext-pdo": "*",
         "ext-zlib": "*",
